+++ packages/react/src/components/FlexBox/index.ts
@@ -1,1 +1,2 @@
 export { default as FlexBox } from './flexBox';
+export * from './FlexBox.types';
